using new clang generator
