+++ Project.toml
@@ -4,14 +4,15 @@ authors = ["Yupei Qi <qiyupei@gmail.com>"]
 version = "2.31.1"
 
 [deps]
+CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
 CSyntax = "ea656a56-6ca6-5dda-bba5-7b6963a5f74c"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 librealsense_jll = "65edd895-dc1e-518e-896c-dc343a3c7f2e"
 
 [compat]
+CSyntax = "0.3.0"
 julia = "1.3"
 librealsense_jll = "2.31.1"
-CSyntax = "0.3.0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
